feat:拆分首屏邮件埋点
